[UPDATE] application - admin integrate 부분 수정

+++ admin/src/main/java/kr/hs/entrydsm/admin/domain/entity/Applicant.java
@@ -57,7 +57,7 @@ public class Applicant {
     private String postCode;
 
     //검정고시 O
-    private Integer averageScore;
+    private BigDecimal averageScore;
 
     //검정고시 X
     private String schoolName;
